requirements.txt
chromadb
streamlit
groq
python-dotenv
langchain
langchain-community
pypdf